# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Api Documentation UMKM Apps",
        "termsOfService": "http://swagger.io/terms/",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        
        "/api/umkm": {
            "get": {
                "tags": [
                    "UMKM"
                ],
                "summary": "Get List UMKM data",
                "description": "Get All List Of UMKM",
                "operationId": "umkm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "S"  
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/umkm/{id}": {
            "get": {
                "tags": [
                    "UMKM"
                ],
                "summary": "Get detail UMKM",
                "description": "Get Detail UMKM",
                "operationId": "detail_umkm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "S"  
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/umkm_products/{id}": {
            "get": {
                "tags": [
                    "UMKM"
                ],
                "summary": "Get List UMKM With Products data",
                "description": "Get All List Of UMKM With Products data",
                "operationId": "umkm_products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "S"  
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/umkm_terdekat": {
            "post": {
                "tags": [
                    "PRODUCT"
                ],
                "summary": "Get List UMKM Terdekat",
                "description": "Get All List Of UMKM Terdekat",
                "operationId": "umkm_terdekat",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "double"
                        }
                    },
                    {
                        "name": "long",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "UMKM"
                ],
                "summary": "Get Detail Product",
                "description": "Get Detail Product",
                "operationId": "product_detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "S"  
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
       
    }
}